+++ Backend/backend-user/src/main/java/com/xiyanchenghong/backenduser/controller/JoinCourseRequest.java
@@ -1,14 +1,13 @@
 package com.xiyanchenghong.backenduser.controller;
 
 public class JoinCourseRequest {
-    private Integer courseId;
+    private Long courseId;
 
-    // Getters and Setters
-    public Integer getCourseId() {
+    public Long getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Integer courseId) {
+    public void setCourseId(Long courseId) {
         this.courseId = courseId;
     }
 }
