+++ Backend/backend-user/src/main/java/com/xiyanchenghong/backenduser/domain/CourseStudent.java
@@ -7,26 +7,26 @@ import java.util.Date;
 public class CourseStudent {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private Integer courseId;
+    private Long id;
+    private Long courseId;
     private Long studentId;
     private Date joinDate;
 
     // Getters and Setters
-    public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCourseId() {
+    public Long getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Integer courseId) {
+    public void setCourseId(Long courseId) {
         this.courseId = courseId;
     }
