+++ Backend/backend-user/src/main/java/com/xiyanchenghong/backenduser/repository/CourseStudentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseStudentRepository extends JpaRepository<CourseStudent, Integer> {
-    CourseStudent findByStudentIdAndCourseId(Long studentId, Integer courseId);
+public interface CourseStudentRepository extends JpaRepository<CourseStudent, Long> {
+    CourseStudent findByStudentIdAndCourseId(Long studentId, Long courseId);
 }
