+++ Backend/backend-user/src/main/java/com/xiyanchenghong/backenduser/service/CourseService.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface CourseService {
     List<Course> getAllCourses();
-    boolean joinCourse(Long userId, Integer courseId);
-    Course getCourseById(Integer courseId);
-    boolean dropCourse(Long userId, Integer courseId);
+    boolean joinCourse(Long userId, Long courseId);
+    Course getCourseById(Long courseId);
+    boolean dropCourse(Long userId, Long courseId);
 }
